+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented here. The format is based
 - Updated underlying OpenSim library version v4.3 (#192)
 - Added "OpenSim" muscle coloring option, which uses `GeometryPath::getColor` (#193)
 - Added ImPlot demo screen to the experimental screens section (#194)
+- Added ability to open multiple muscle plots in the editor (related: #191)
 
 
 ## [0.1.5] - 2022/04/11